requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.36.0
pandas>=2.0.0
matplotlib>=3.7.0
wordcloud>=1.9.0
scikit-learn>=1.3.0
nltk>=3.8.0
transformers>=4.40.0
torch>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
altair>=5.0.0
